refactor: rename nwd folder and combine imports
