+++ src/index.js
@@ -1,9 +1,7 @@
 import readline from 'node:readline';
 import process from 'node:process';
 
-import { up } from './commands/up.js';
-import { cd } from './commands/cd.js';
-import { ls } from './commands/ls.js';
+import { up, cd, ls } from './nwd/index.js';
 
 const DEFAULT_USERNAME = 'User';
 const EXIT_CMD = '.exit';
