+++ src/nwd/index.js
@@ -1,0 +1,5 @@
+import { up } from './up.js';
+import { cd } from './cd.js';
+import { ls } from './ls.js';
+
+export { up, cd, ls };
